update
Add the second CPU for rendering, added an affine voxel drawing function and added dithered alpha rendering for polygons and sprites that can be used for fog or water among other fixes

+++ shared/source/engine_globals.h
@@ -30,7 +30,7 @@ extern s16 ordering_table[SCN_OT_SIZE];
 extern pl_list_t ot_pl_list[SCN_VIS_PL_LIST_SIZE];
 extern u16 dl_pl_list[SCN_VIS_PL_LIST_SIZE], dl_vt_list[SCN_VIS_VT_LIST_SIZE], dl_obj_list[SCN_MAX_NUM_OBJECTS], dl_tr_vt_index[MDL_VT_LIST_MAX_SIZE];
 #if ENABLE_TEXTURE_BUFFER
-  extern u8 texture_buffer[TEXTURE_BUFFER_SIZE];
+  extern u16 texture_buffer[TEXTURE_BUFFER_SIZE];
 #endif
 extern int profile_vars[3];
 // extern const u8 *texture_image;
